Consolidate SSH functions in xssh package
Originally the plan was to have a verifier package that would provide
various verification functions (e.g. SSH, PGP, etc., with SSH to start
with). But we didn't anticipate having to write our own allowed_signers
parser and so on, so it makes more sense to bundle functionality by
signature type. We name it xssh so as not to clash with x/crypto's ssh,
which we are extending.