Refactor configuration
The intention is to move configuration and variable loading as far as
possible from the service, eventually to improve testability but also to
abstract the way we load configuration.

We also move loading allowed signers from an io.Reader to xssh, since it
seems thematically appropriate.

We're left with a relatively lean service package that focuses more on
serving now that we moved the setup out.